requirements.txt
# =============================================================================
# DFToolBench-I — requirements.txt
# =============================================================================
# Usage:
#   pip install -r requirements.txt            # core dependencies
#   pip install -r requirements.txt -r requirements-eval.txt   # + LLM eval
#
# For tool-specific conda envs see: envs/
# =============================================================================

# -----------------------------------------------------------------------------
# Deep learning core
# -----------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# -----------------------------------------------------------------------------
# Vision & image I/O
# -----------------------------------------------------------------------------
Pillow>=9.0.0
opencv-python>=4.7.0
imageio>=2.28.0

# -----------------------------------------------------------------------------
# Numerical / scientific
# -----------------------------------------------------------------------------
numpy>=1.23.0
scipy>=1.10.0
scikit-image>=0.21.0

# -----------------------------------------------------------------------------
# OCR (OCRTool)
# -----------------------------------------------------------------------------
easyocr>=1.7.0

# -----------------------------------------------------------------------------
# Transformers & vision models
# (used by DeepfakeDetectionTool, AnomalyDetectionTool, SegmentationTool)
# -----------------------------------------------------------------------------
transformers>=4.38.0
timm>=0.9.0
einops>=0.6.0
huggingface-hub>=0.22.0

# CLIP backbone used by D3 (DeepfakeDetectionTool)
open-clip-torch>=2.20.0

# -----------------------------------------------------------------------------
# Object detection (ObjectDetectionTool — YOLOE/YOLOv8)
# -----------------------------------------------------------------------------
ultralytics>=8.0.0

# -----------------------------------------------------------------------------
# HTTP tool server & bridge
# -----------------------------------------------------------------------------
flask>=3.0.0
requests>=2.31.0
gunicorn>=21.0.0
werkzeug>=3.0.0

# -----------------------------------------------------------------------------
# Data handling & serialisation
# -----------------------------------------------------------------------------
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
jsonschema>=4.17.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
rich>=13.0.0
click>=8.1.0

# =============================================================================
# Optional — LLM evaluation clients
# Install with:  pip install "dftoolbench[eval]"  or uncomment below.
# =============================================================================

# OpenAI (GPT-4, GPT-4o, GPT-5 …)
# openai>=1.30.0
# tiktoken>=0.6.0

# Anthropic (Claude series)
# anthropic>=0.25.0

# Google (Gemini series)
# google-generativeai>=0.5.0

# Sentence similarity (SummAcc judge)
# sentence-transformers>=2.7.0

# OpenCompass evaluation framework
# opencompass>=0.2.0

# =============================================================================
# Optional — local / open-weight LLM inference
# =============================================================================

# vLLM for efficient local inference
# vllm>=0.4.0
# accelerate>=0.29.0
# bitsandbytes>=0.43.0

# =============================================================================
# Optional — development & testing
# =============================================================================

# pytest>=8.0.0
# pytest-cov>=5.0.0
# black>=24.0.0
# isort>=5.13.0
# flake8>=7.0.0
# mypy>=1.9.0
# pre-commit>=3.7.0
